# doc/api/docextract-api-docs.json
{"openapi":"3.1.0","info":{"title":"DocExtract API","version":"v1"},"servers":[{"url":"http://localhost:8080/adeon-docextract","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/value-lists/{propertyId}":{"post":{"tags":["retrieval-controller"],"operationId":"valueList","parameters":[{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValueListRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DmsValueList"}}}}}}},"/api/v1/validation/reject":{"post":{"tags":["validation-controller"],"operationId":"reject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/validation/confirm":{"post":{"tags":["validation-controller"],"operationId":"confirm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConfirmationResult"}}}}}}},"/api/v1/ingest/documents":{"post":{"tags":["upload-controller"],"operationId":"upload","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IngestedDocument"}}}}}}},"/api/v1/processes/{processId}/similar-documents":{"get":{"tags":["retrieval-controller"],"operationId":"similarDocuments","parameters":[{"name":"processId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RetrievalResult"}}}}}}}},"/api/v1/processes/{processId}/extraction-result":{"get":{"tags":["extraction-controller"],"operationId":"extractionResult","parameters":[{"name":"processId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExtractionResult"}}}}}}},"/api/v1/processes/{processId}/events":{"get":{"tags":["process-events-controller"],"operationId":"subscribe","parameters":[{"name":"processId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/v1/processes/{processId}/document-type-candidates":{"get":{"tags":["retrieval-controller"],"operationId":"documentTypeCandidates","parameters":[{"name":"processId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DmsDocumentMetadata"}}}}}}}},"/api/v1/object-definitions":{"get":{"tags":["retrieval-controller"],"operationId":"objectDefinitions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DmsDocumentMetadata"}}}}}}}},"/api/v1/ingest/documents/{blobId}/preview":{"get":{"tags":["preview-controller"],"operationId":"preview","parameters":[{"name":"blobId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"Range","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/v1/ingest/documents/test":{"get":{"tags":["upload-controller"],"operationId":"getMethodName","parameters":[{"name":"param","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ValueListRequest":{"type":"object","properties":{"objectDefinitionId":{"type":"string"},"extendedProperties":{"type":"object","additionalProperties":{"type":"string"}},"multivalueExtendedProperties":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}},"searchTerm":{"type":"string"}}},"DmsValueList":{"type":"object","properties":{"values":{"type":"array","items":{"type":"string"}},"hasMore":{"type":"boolean"}}},"RejectRequest":{"type":"object","properties":{"processId":{"type":"string"}}},"ConfirmRequest":{"type":"object","properties":{"processId":{"type":"string"},"documentId":{"type":"string"},"repositoryId":{"type":"string"},"consentGiven":{"type":"boolean"},"templateConsent":{"type":"boolean"},"attributes":{"type":"object","additionalProperties":{}}}},"ConfirmationResult":{"type":"object","properties":{"documentId":{"type":"string"},"writtenBack":{"type":"boolean"},"templateConsentRecorded":{"type":"boolean"}}},"IngestedDocument":{"type":"object","properties":{"processId":{"type":"string"},"dmsLocation":{"type":"string"},"originalBlobId":{"type":"string"},"mediaType":{"type":"string"}}},"RetrievalResult":{"type":"object","properties":{"repositoryId":{"type":"string"},"documentId":{"type":"string"},"score":{"type":"number","format":"double"},"properties":{"type":"string"}}},"ExtractedAttribute":{"type":"object","properties":{"propertyId":{"type":"string"},"value":{"type":"string"},"values":{"type":"array","items":{"type":"string"}},"confidence":{"type":"number","format":"double"},"sourceExcerpt":{"type":"string"}}},"ExtractionResult":{"type":"object","properties":{"documentTypeId":{"type":"string"},"attributes":{"type":"array","items":{"$ref":"#/components/schemas/ExtractedAttribute"}}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"DmsDocumentMetadata":{"type":"object","properties":{"documentType":{"$ref":"#/components/schemas/DmsDocumentType"},"contentLanguage":{"type":"string"},"properties":{"type":"array","items":{"$ref":"#/components/schemas/DmsProperty"}}}},"DmsDocumentType":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"properties":{"type":"array","items":{"$ref":"#/components/schemas/DmsPropertyDefinition"}}}},"DmsProperty":{"type":"object","properties":{"id":{"type":"string"},"value":{"type":"string"},"values":{"type":"array","items":{"type":"string"}}}},"DmsPropertyDefinition":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"dataType":{"type":"string","enum":["ALPHANUMERIC","NUMERIC","MONEY","DATETIME","DATE","KEY_VALUE_LIST","COLOR_CODE","DOCUMENT_TYPE","UNKNOWN"]},"mandatory":{"type":"boolean"},"hasValueList":{"type":"boolean"},"isDynamicValueList":{"type":"boolean"},"isHidden":{"type":"boolean"},"isModifiable":{"type":"boolean"},"isSystemProperty":{"type":"boolean"},"isMultiAttribute":{"type":"boolean"},"isDocumentTypeProperty":{"type":"boolean"},"maxRowCount":{"type":"integer","format":"int32"},"maxLength":{"type":"integer","format":"int32"},"valueList":{"type":"array","items":{"type":"string"}},"hasMoreValues":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}